requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
pymongo
nltk
spacy
python-multipart
python-docx
pytesseract
Pillow
pdf2image
pdfminer.six
fastapi
uvicorn
python-multipart
nltk
spacy
pytesseract
Pillow

